- full db persistence

+++ lib/database.dart
@@ -11,21 +11,11 @@ var goalsFixture = GoalsModel.from(
     Goal(name: 'Learn BWV772 ?', goalType: GoalType.learning, tasks: [
       Task(name: 'LH Reading', estimation: Duration(minutes: 10)),
       Task(name: 'RH Reading', estimation: Duration(minutes: 10)),
-      Task(
-          name: 'Practice C Scale',
-          estimation: Duration(minutes: 10)),
-      Task(
-          name: 'Practice D Scale',
-          estimation: Duration(minutes: 10)),
-      Task(
-          name: 'Practice A minor Scale',
-          estimation: Duration(minutes: 10)),
-      Task(
-          name: 'Measures 1-4',
-          estimation: Duration(minutes: 10)),
-      Task(
-          name: 'Measures 2-8',
-          estimation: Duration(minutes: 10)),
+      Task(name: 'Practice C Scale', estimation: Duration(minutes: 10)),
+      Task(name: 'Practice D Scale', estimation: Duration(minutes: 10)),
+      Task(name: 'Practice A minor Scale', estimation: Duration(minutes: 10)),
+      Task(name: 'Measures 1-4', estimation: Duration(minutes: 10)),
+      Task(name: 'Measures 2-8', estimation: Duration(minutes: 10)),
     ]),
     Goal(
         name: 'Read Good Strategy Bad Strategy',
@@ -35,26 +25,20 @@ var goalsFixture = GoalsModel.from(
           Task(name: 'Chapter 2', estimation: Duration(minutes: 15)),
         ]),
     Goal(name: 'Bake a traditional pie', goalType: GoalType.learning, tasks: [
-      Task(
-          name: 'Buy ingredients', estimation: Duration(minutes: 60)),
-      Task(
-          name: 'Mix all together',
-          estimation: Duration(minutes: 15)),
+      Task(name: 'Buy ingredients', estimation: Duration(minutes: 60)),
+      Task(name: 'Mix all together', estimation: Duration(minutes: 15)),
       Task(name: 'Bake', estimation: Duration(minutes: 30)),
     ]),
   ],
 );
 
 class AppDatabaseMigrations {
-  static const goalsTable = 'goals';
-  static const tasksTable = 'tasks';
-
   void v1up(Database db) async {
     var batch = db.batch();
 
     batch.execute(
       '''
-      CREATE TABLE $goalsTable(
+      CREATE TABLE ${AppDatabase.goalsTable}(
         id INTEGER PRIMARY KEY, 
         name TEXT,
         type TEXT
@@ -64,7 +48,7 @@ class AppDatabaseMigrations {
 
     batch.execute(
       '''
-      CREATE TABLE $tasksTable(
+      CREATE TABLE ${AppDatabase.tasksTable}(
         id INTEGER PRIMARY KEY,         
         name TEXT,
         estimation INTEGER,
@@ -77,13 +61,12 @@ class AppDatabaseMigrations {
   }
 
   void addFixture(Database db) async {
-    await db.delete(goalsTable);
-
+    await db.delete(AppDatabase.goalsTable);
 
     var batch = db.batch();
     for (var i = 0; i < goalsFixture.items.length; i++) {
       var goal = goalsFixture.items[i];
-      batch.insert(goalsTable, goal.toMap());
+      batch.insert(AppDatabase.goalsTable, goal.toMap());
     }
     var ids = await batch.commit();
 
@@ -91,22 +74,20 @@ class AppDatabaseMigrations {
       goalsFixture.items[i].id = ids[i] as int;
     }
 
-
     for (var i = 0; i < goalsFixture.items.length; i++) {
       var goal = goalsFixture.items[i];
 
       batch = db.batch();
-      for (var j=0;j<goal.tasks.length; j++) {
+      for (var j = 0; j < goal.tasks.length; j++) {
         goal.tasks[j].goalId = goal.id;
-        batch.insert(tasksTable, goal.tasks[j].toMap());
+        batch.insert(AppDatabase.tasksTable, goal.tasks[j].toMap());
       }
       ids = await batch.commit();
 
-      for (var j=0;j<goal.tasks.length; j++) {
+      for (var j = 0; j < goal.tasks.length; j++) {
         goal.tasks[j].id = ids[j] as int;
       }
     }
-
 
     goalsFixture.items.forEach((goal) {
       log('now $goal');
@@ -115,6 +96,9 @@ class AppDatabaseMigrations {
 }
 
 class AppDatabase {
+  static const goalsTable = 'goals';
+  static const tasksTable = 'tasks';
+
   static AppDatabase? _instance;
 
   AppDatabase._privateConstructor(this._database);
@@ -124,8 +108,9 @@ class AppDatabase {
   static Future<Database> _init() async {
     String dbName = join(await getDatabasesPath(), 'taskchisel.db');
 
-    await databaseExists(dbName)
-        .then((exists) => exists ? deleteDatabase(dbName) : null);
+    if (false) // enable to drop the db and rebuild it
+      await databaseExists(dbName)
+          .then((exists) => exists ? deleteDatabase(dbName) : null);
 
     final database = await openDatabase(
       dbName,
@@ -146,7 +131,7 @@ class AppDatabase {
     return _instance!;
   }
 
-  Database get database =>  _database;
+  Database get database => _database;
 
   static void _migrate(db, version) async {
     var _migrations = AppDatabaseMigrations();
